% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{adjust_qc_means}
\alias{adjust_qc_means}
\title{Adjust QC Means}
\usage{
adjust_qc_means(FUNC_list, master_list)
}
\arguments{
\item{FUNC_list}{A list containing the master data and corrected data.}

\item{master_list}{A list containing the project details and data.}
}
\value{
The updated \code{FUNC_list} with adjusted QC means in the corrected data.
}
\description{
This function adjusts the means of QC samples in the corrected data by calculating the correction ratio based on original and corrected means.
It applies the correction ratio to the corrected data and updates the sample type for QC samples.
}
