% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{apply_lgw_imputation}
\alias{apply_lgw_imputation}
\title{Apply LGW Imputation}
\usage{
apply_lgw_imputation(mat)
}
\arguments{
\item{mat}{dataframe or matrix for imputation to be applied}
}
\description{
A wrapper to apply lgw_impute to the dataframe
}
\keyword{internal}
