% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{archive_raw_files}
\alias{archive_raw_files}
\title{Archive Raw Files}
\usage{
archive_raw_files(project_directory)
}
\arguments{
\item{project_directory}{Path to the directory for the project parsed from SkylineR.}
}
\value{
None. The function performs the archive operation and a message upon successful completion.
}
\description{
This function moves raw files (wiff and mzML) to an archive directory after processing is complete.
}
\examples{
\dontrun{
archive_raw_files("path/to/project_directory")
}
}
