% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{assign_sample_type}
\alias{assign_sample_type}
\title{Assign Sample Type}
\usage{
assign_sample_type(sample_tags, run_order)
}
\arguments{
\item{sample_tags}{A character vector passed from master_list$project_details$sample_tags, containing QC types.}

\item{run_order}{A data frame containing the run order information.}
}
\description{
This function assigns sample types based on the sample names and user-defined QC types.
}
\keyword{internal}
