% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{calculate_baseline}
\alias{calculate_baseline}
\title{calculate_baseline}
\usage{
calculate_baseline(FUNC_mzR, idx_plate, idx_mzML, idx_mrm)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{idx_plate}{A string naming the current plate ID being processed. Passed from process_files function.}

\item{idx_mzML}{A string naming the current mzML file being processed. Passed from process_files function.}

\item{idx_mrm}{A string identifying the current mrm. Passed from process_files function.}
}
\value{
A numeric value for baseline.
}
\description{
This function calculates baseline for a single transition.
}
\examples{
\dontrun{
calculate_baseline(FUNC_mzR, idx_plate, idx_mzML, idx_mrm)
}
}
\keyword{internal}
