% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{calculate_total_runtime}
\alias{calculate_total_runtime}
\title{Calculate Total Runtime}
\usage{
calculate_total_runtime(master_list, section_name)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{section_name}{A string representing the name of the current section.}
}
\value{
None
}
\description{
This function calculates the total runtime from the start.
}
\keyword{internal}
