% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{capture_current_time}
\alias{capture_current_time}
\title{Capture Current Time}
\usage{
capture_current_time(master_list, section_name)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{section_name}{A string representing the name of the current section.}
}
\value{
updated master list
}
\description{
This function captures the current time for the given section.
}
\keyword{internal}
