% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{check_and_prompt_suggested_packages}
\alias{check_and_prompt_suggested_packages}
\title{Check and Prompt Suggested Packages}
\usage{
check_and_prompt_suggested_packages(description_path = "DESCRIPTION")
}
\arguments{
\item{description_path}{Character. MetaboExploreR DESCRIPTION file. Default is "DESCRIPTION".}
}
\value{
Invisible TRUE if all suggested packages are installed, FALSE otherwise prompts
user to install missing packages
}
\description{
This function reads the DESCRIPTION file of MetaboExploreR, checks for suggested packages,
and prompts the user to install any missing packages.
}
\examples{
\dontrun{
check_and_prompt_suggested_packages()
}
}
