% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{check_sil_standards}
\alias{check_sil_standards}
\title{check_sil_standards}
\usage{
check_sil_standards(master_list, plate_idx, version)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to check the SIL standards for.}

\item{version}{The version of the MRM guide to compare against.}
}
\value{
A logical value indicating whether the SIL standards on the plate match the SIL standards for the specified version.
}
\description{
This function checks if the SIL standards on a given plate match the SIL standards for a specified mrm_guide version.
}
\examples{
\dontrun{
check_sil_standards(master_list, plate_idx, version)
}
}
