% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{clean_statTarget_output}
\alias{clean_statTarget_output}
\title{Clean statTarget Output}
\usage{
clean_statTarget_output(data)
}
\arguments{
\item{data}{A tibble containing the output data from \code{statTarget}.}
}
\value{
A cleaned tibble with renamed columns and numeric data types.
}
\description{
This function cleans the output data from \code{statTarget} by filtering out unwanted rows and renaming columns.
It handles different column structures based on the presence of specific sample columns.
}
\keyword{internal}
