% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionCheckR.R
\name{compare_mrm_template_with_guide}
\alias{compare_mrm_template_with_guide}
\title{compare_mrm_template_with_guide}
\usage{
compare_mrm_template_with_guide(mrm_template, concentration_guide)
}
\arguments{
\item{mrm_template}{df of transitions (mrm_template) for SkylineR or qcCheckR}

\item{concentration_guide}{df of concentrations for SIL internal standards}
}
\value{
A message if all transitions are unique or a list of non unique transitions to be fixed
}
\description{
This function checks if all internal standards from the note column in the transition list has a match in concentration guide
}
\examples{
\dontrun{
compare_mrm_template_with_guide(mrm_template, concentration_guide)
}
}
