% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{construct_command_for_terminal}
\alias{construct_command_for_terminal}
\title{construct_command_for_terminal}
\usage{
construct_command_for_terminal(wiff_file_paths, project_directory)
}
\arguments{
\item{wiff_file_paths}{Vector of wiff file paths.}

\item{project_directory}{Directory path for the project folder.}
}
\value{
The constructed command string.
}
\description{
This function constructs the command for terminal to convert wiff files to mzML format.
}
\examples{
\dontrun{
command <- construct_command_for_terminal(wiff_file_paths, "path/to/project_directory")
}
}
