% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{copy_files}
\alias{copy_files}
\title{copy_files
This function copies files from the source directory to the destination directory.}
\usage{
copy_files(source_dir, dest_dir)
}
\arguments{
\item{source_dir}{Directory path for the folder to copy.}

\item{dest_dir}{Directory path for the folder to be copied to.}
}
\value{
A vector of file paths that were copied.
}
\description{
copy_files
This function copies files from the source directory to the destination directory.
}
\examples{
\dontrun{
copy_files(source_dir = "path/to/source", dest_dir = "path/to/destination")
}
}
\keyword{internal}
