% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_message}
\alias{create_message}
\title{Create Message}
\usage{
create_message(master_list, section_name, next_section_name)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{section_name}{A string representing the name of the current section.}

\item{next_section_name}{A string representing the name of the next section.}
}
\value{
None
}
\description{
This function creates a message for the log.
}
