% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{create_output}
\alias{create_output}
\title{create_output}
\usage{
create_output(FUNC_tibble, FUNC_mrm_guide, mzML_filelist)
}
\arguments{
\item{FUNC_tibble}{A tibble of mrm details per sample per lipid species.}

\item{FUNC_mrm_guide}{Tibble of mrm details parsed internally. See run mrm_template_guide for example.}

\item{mzML_filelist}{A list containing mzML file information from plate.}
}
\value{
A list containing two tibbles for mrm_guide_updated and peak_boundary_update.
}
\description{
This function updated mrm_guide and peak_boundary_update for later use in Skyline.
}
\examples{
\dontrun{
create_output(FUNC_tibble, FUNC_mrm_guide, mzML_filelist)
}
}
