% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{create_pheno_file}
\alias{create_pheno_file}
\title{Create Pheno File}
\usage{
create_pheno_file(FUNC_list)
}
\arguments{
\item{FUNC_list}{A list containing the master data and project directory.}
}
\value{
The updated \code{FUNC_list} with the created phenotype file.
}
\description{
This function creates a phenotype file from the master data.
It selects relevant columns, renames them, and formats the sample IDs and classes.
}
