% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{create_profile_file}
\alias{create_profile_file}
\title{Create Profile File}
\usage{
create_profile_file(FUNC_list)
}
\arguments{
\item{FUNC_list}{A list containing the master data, metabolite list, and project directory.}
}
\value{
The updated \code{FUNC_list} with the created profile file.
}
\description{
This function creates a profile file from the master data.
It selects the sample names and metabolite data, renames columns, and formats the data into a matrix.
}
\keyword{internal}
