% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{create_summary_table}
\alias{create_summary_table}
\title{create_summary_table}
\usage{
create_summary_table(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to create the summary table for.}
}
\value{
A tibble containing the summary table with project details and their values.
}
\description{
This function creates a summary table for a given plate in the master list, including various project details.
}
\examples{
\dontrun{
create_summary_table(master_list, plate_idx)
}
}
\keyword{internal}
