% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{create_user_guide}
\alias{create_user_guide}
\title{Create User Guide}
\usage{
create_user_guide(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
A tibble containing the user guide with key metrics and descriptions.
}
\description{
This function creates a user guide for the \code{master_list} project, summarizing key project details and metrics.
}
\keyword{internal}
