% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{determine_qc_type}
\alias{determine_qc_type}
\title{Determine QC Type}
\usage{
determine_qc_type(global_qc_pass)
}
\arguments{
\item{global_qc_pass}{A list containing the global QC pass status.}
}
\value{
A string indicating the QC type ("pqc", "ltr", or "unknown").
}
\description{
This function determines the QC type based on the global QC pass status.
It checks if the pqc or ltr QC types have passed and returns the appropriate type.
}
