% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{execute_command}
\alias{execute_command}
\title{execute_command}
\usage{
execute_command(command)
}
\arguments{
\item{command}{The command string to execute.}
}
\value{
None. The function executes the command.
}
\description{
This function executes the command to convert files to mzML format.
}
\examples{
\dontrun{
execute_command(command)
}
}
