% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{export_files}
\alias{export_files}
\title{export_files}
\usage{
export_files(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to export files for.}
}
\value{
Exports CSV, SKY, and TSV files to the specified project directory.
}
\description{
This function exports various files related to the project for a given plate, including updated MRM guides, peak boundaries, and default templates.
}
\examples{
\dontrun{
export_files(master_list, plate_idx)
}
}
\keyword{internal}
