% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{export_master_list_rda}
\alias{export_master_list_rda}
\title{Export Master List as RDA File
This function exports the \code{master_list} to an RDA file, which can be used for future analysis or sharing.}
\usage{
export_master_list_rda(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the RDA file exported.
}
\description{
Export Master List as RDA File
This function exports the \code{master_list} to an RDA file, which can be used for future analysis or sharing.
}
