% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{export_xlsx_file}
\alias{export_xlsx_file}
\title{Export XLSX File}
\usage{
export_xlsx_file(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the XLSX file exported.
}
\description{
This function exports the \code{master_list} data to an XLSX file, including user
guide, QC metrics, and processed data.
}
\keyword{internal}
