% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{extract_timestamp}
\alias{extract_timestamp}
\title{extract_timestamp}
\usage{
extract_timestamp(timestamp)
}
\arguments{
\item{timestamp}{A string representing the timestamp.}
}
\value{
A numeric value representing the year extracted from the timestamp.
}
\description{
This function extracts the year from a timestamp string and converts it to a numeric value.
}
\examples{
\dontrun{
extract_timestamp("2025-06-17T13:31:58Z")
}
}
