% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{filter_concentration}
\alias{filter_concentration}
\title{Filter Concentration Data}
\usage{
filter_concentration(master_list, source)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{source}{The data source to filter (e.g., "concentration", "concentration\link{statTarget}").}
}
\value{
A tibble containing the filtered concentration data.
}
\description{
This function filters the concentration data from the \code{master_list} based on the specified source.
It removes failed samples and lipids, and applies RSD filters based on the specified source.
}
