% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{finalise_sorted_data}
\alias{finalise_sorted_data}
\title{Finalise Sorted Data}
\usage{
finalise_sorted_data(master_list)
}
\arguments{
\item{master_list}{A list containing project details and sorted peak area data.}
}
\value{
The updated master list with finalised sorted data.
}
\description{
This function finalises the sorted data by adding sample type factors, reversing the order of sample types, and setting the sample data source.
}
\keyword{internal}
