% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{find_method_version}
\alias{find_method_version}
\title{Find method versions}
\usage{
find_method_version(master_list)
}
\arguments{
\item{master_list}{A list containing porject details and data.}
}
\value{
The updated \code{master_list} with method versions found for each plate.
}
\description{
This function finds the method versions for each plate in the master list.
It matches internal standards to the SIL guide and stores the matching SIL_guide.
}
