% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{find_peak_end_idx}
\alias{find_peak_end_idx}
\title{find_peak_end_idx}
\usage{
find_peak_end_idx(
  FUNC_mzR,
  idx_plate,
  idx_mzML,
  idx_mrm,
  peak_apex_idx,
  baseline_value
)
}
\arguments{
\item{FUNC_mzR}{List from master_list containing mzR object for each sample, mzR_header, mzR_chromatogram parsed internally.}

\item{idx_plate}{A string naming the current plate ID being processed. Passed from process_files function.}

\item{idx_mzML}{A string naming the current mzML file being processed. Passed from process_files function.}

\item{idx_mrm}{A string identifying the current mrm. Passed from process_files function.}

\item{peak_apex_idx}{Index of peak apex passed from process_files function.}

\item{baseline_value}{Numeric value indicating transitions baseline. Passed from process_files function.}
}
\value{
A numeric value for peak end index.
}
\description{
This function finds peak end for a single transition.
}
\examples{
\dontrun{
find_peak_end_idx(FUNC_mzR, idx_plate, idx_mzML, idx_mrm, peak_apex_idx, baseline_value)
}
}
\keyword{internal}
