% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{flag_failed_qc_injections}
\alias{flag_failed_qc_injections}
\title{Flag Failed QC Injections}
\usage{
flag_failed_qc_injections(FUNC_list)
}
\arguments{
\item{FUNC_list}{A list containing the master data and metabolite list.}
}
\value{
The updated \code{FUNC_list} with flagged failed QC injections.
}
\description{
This function flags failed QC injections by checking the signal intensity of QC samples.
It identifies QC samples with low signal intensity and marks them as "sample" in the master data.
}
