% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{format_rsd_table}
\alias{format_rsd_table}
\title{Format RSD Table}
\usage{
format_rsd_table(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
A tibble containing the formatted RSD table with rounded values and transposed structure.
}
\description{
This function formats the RSD table from the \code{master_list} filters.
It rounds the RSD values, adds a data column, and transposes the table for better readability.
}
\keyword{internal}
