% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{generate_inter_plate_summary}
\alias{generate_inter_plate_summary}
\title{Generate Inter-Plate Summary}
\usage{
generate_inter_plate_summary(master_list, metrics, sample_tags)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{metrics}{tibble of the metrics to be included in the summary.}

\item{sample_tags}{A vector of sample type tags to be included in the summary.}
}
\value{
A tibble containing the inter-plate summary metrics.
}
\description{
This function generates a summary of all plates in the \code{master_list}.
It aggregates data across all plates, including metrics such as matrix type, sample counts, lipid targets, SIL versions, missing value filter flags, and RSD percentages.
}
