% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{get_common_control_metabolites}
\alias{get_common_control_metabolites}
\title{Get Common Control Metabolites
This function retrieves the common control metabolites across different SIL versions in the \code{master_list}.
It intersects the precursor names of control charts from each SIL version's guide.}
\usage{
get_common_control_metabolites(master_list)
}
\arguments{
\item{master_list}{A list containing project details and templates.}
}
\value{
A vector of common control metabolites.
}
\description{
Get Common Control Metabolites
This function retrieves the common control metabolites across different SIL versions in the \code{master_list}.
It intersects the precursor names of control charts from each SIL version's guide.
}
