% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{get_plate_annotations}
\alias{get_plate_annotations}
\title{Get Plate Annotations}
\usage{
get_plate_annotations(master_list)
}
\arguments{
\item{master_list}{A list containing project details and PCA scores.}
}
\value{
A tibble containing plate annotations with median run indices.
}
\description{
This function retrieves the median run index for each plate in the \code{master_list} data.
It creates a tibble with sample data source, plate ID, run index, and placeholder values for PCA components and value.
}
