% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{import_mzml}
\alias{import_mzml}
\title{import_mzml}
\usage{
import_mzml(plateID, master_list)
}
\arguments{
\item{plateID}{Plate ID for the current plate.}

\item{master_list}{Master list generated internally.}
}
\value{
The updated \code{master_list} object with the mzML import details.
}
\description{
This function imports mzML files for each plate using the mzR package, extracts relevant information, and updates the script log.
}
\examples{
\dontrun{
import_mzml("plateID", master_list)
}
}
