% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{initialise_lipid_filter}
\alias{initialise_lipid_filter}
\title{Initialise lipid filter structure}
\usage{
initialise_lipid_filter(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with initialised lipid filter structure.
}
\description{
This function initialses the structure for lipid filtering in the \code{master_list}.
It creates a list to store lipid missing values, a summary data frame, and lists for project flags and failed lipids.
}
\keyword{internal}
