% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{initialise_master_list}
\alias{initialise_master_list}
\title{initialise_master_list}
\usage{
initialise_master_list()
}
\value{
A list representing the initialised master list.
}
\description{
This function initialises the master list with default values.
}
\examples{
\dontrun{
master_list <- initialise_master_list()
}
}
