% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{initialise_mzml_filelist}
\alias{initialise_mzml_filelist}
\title{initialise_mzml_filelist}
\usage{
initialise_mzml_filelist(master_list)
}
\arguments{
\item{master_list}{A list containing project details, including plate IDs and project directory.}
}
\value{
A list of mzML files for each plate, excluding files with "COND", "Blank", or "ISTDs" in their names.
}
\description{
This function initializes a list of mzML files for each plate in the master list, excluding files with specific patterns.
}
\examples{
\dontrun{
initialise_mzml_filelist(master_list)
}
}
\keyword{internal}
