% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{initialise_sil_summary}
\alias{initialise_sil_summary}
\title{Initialise SIL Summary Data Frame}
\usage{
initialise_sil_summary()
}
\value{
An empty data frame with specified columns for SIL summary statistics.
}
\description{
This function initializes an empty data frame to store SIL summary statistics.
It includes columns for lipid names, template versions, plate IDs, and various flags related to SIL internal standards.
}
\keyword{internal}
