% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{initialise_statTarget_environment}
\alias{initialise_statTarget_environment}
\title{Initialise statTarget Environment}
\usage{
initialise_statTarget_environment(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
A list containing the project directory, master data, and metabolite list for \code{statTarget}.
}
\description{
This function initialises the environment for \code{statTarget} batch correction.
It creates necessary directories, sets up the master data, and flags failed QC injections.
}
\keyword{internal}
