% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_MetaboExploreR}
\alias{install_MetaboExploreR}
\title{install_MetaboExploreR}
\usage{
install_MetaboExploreR()
}
\description{
Helper function to install MetaboExplorer. Ensuring all packages
are installed and up to date.
}
\examples{
\dontrun{
install_MetaboExploreR()
}
}
