% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{integrate_corrected_data}
\alias{integrate_corrected_data}
\title{Integrate Corrected Data into Master List}
\usage{
integrate_corrected_data(master_list, FUNC_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{FUNC_list}{A list containing the corrected data from \code{statTarget}.}
}
\value{
The updated \code{master_list} with integrated corrected data.
}
\description{
This function integrates the corrected data from \code{statTarget} into the master list.
It splits the corrected data by sample plate ID, updates the sample data source, and processes the peak area and concentration data.
}
