% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{lgw_impute}
\alias{lgw_impute}
\title{LGW Impute}
\usage{
lgw_impute(x)
}
\arguments{
\item{x}{A tibble of peak area concentrations.}
}
\value{
A tibble with imputed values.
}
\description{
This function imputes missing values in a tibble of peak area concentrations by replacing NAs with half the minimum non-zero value.
}
\examples{
\dontrun{
lgw_impute(tibble(a = c(0, 1, 2, NA), b = c(3, 0, NA, 4)))
}
}
