% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Merge_Metadata}
\usage{
merge_metadata(original_data, imputed_data)
}
\arguments{
\item{original_data}{A tibble containing the original peak area data with sample names.}

\item{imputed_data}{A tibble containing the imputed peak area data with sample names.}
}
\value{
A tibble with merged metadata and imputed data, including a sample data source column.
}
\description{
This function merges metadata from the original data with the imputed data.
It selects sample name columns from the original data and performs a left join with the imputed data.
It also adds a column indicating the data source of the imputed data.
}
