% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{move_folder}
\alias{move_folder}
\title{move_folder
This function moves a folder from the source directory to the destination directory, waits until files are not in use, and then deletes the source directory.}
\usage{
move_folder(source_dir, dest_dir, max_wait = 60, max_retries = 30)
}
\arguments{
\item{source_dir}{Directory path for the folder to copy.}

\item{dest_dir}{Directory path for the folder to be copied to.}

\item{max_wait}{Maximum wait time (in seconds) for the system prior to deleting the moved folder.}

\item{max_retries}{Maximum number of attempts to try move/delete prior to error. Default 30}
}
\value{
None. The function performs the move operation and a message upon successful completion.
}
\description{
move_folder
This function moves a folder from the source directory to the destination directory, waits until files are not in use, and then deletes the source directory.
}
\examples{
\dontrun{
move_folder(source_dir = "path/to/source", dest_dir = "path/to/destination", max_wait = 60, max_retries = 30)
}
}
