% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_execute_command}
\alias{msConvertR_execute_command}
\title{msConvertR_execute_command}
\usage{
msConvertR_execute_command(command)
}
\arguments{
\item{command}{The command string to execute.}
}
\value{
None. The function executes the command.
}
\description{
This function executes the command to convert files to mzML format.
}
\examples{
\dontrun{
msConvertR_execute_command(command)
}
}
