% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_mzml_conversion}
\alias{msConvertR_mzml_conversion}
\title{msConvertR_mzml_conversion}
\usage{
msConvertR_mzml_conversion(
  input_directory,
  output_directory,
  plateIDs,
  vendor_extension_patterns
)
}
\arguments{
\item{input_directory}{Directory path for project folder}

\item{output_directory}{Directory path for project folder if different from
input directory.}

\item{plateIDs}{vector of vendor files names to be converted}

\item{vendor_extension_patterns}{character string of vendor file extensions.}
}
\value{
Converted mzml files.
}
\description{
This function converts raw vendor files to mzML format using ProteoWizard's msconvert tool, restructures directories, and updates the script log.
}
\examples{
\dontrun{
msConvertR_mzml_conversion(input_directory, output_directory, plateIDs)
}
}
\keyword{internal}
