% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_restructure_directory}
\alias{msConvertR_restructure_directory}
\title{msConvertR_restructure_directory}
\usage{
msConvertR_restructure_directory(output_directory, plateID)
}
\arguments{
\item{output_directory}{Output directory where the mzML files will be stored.}

\item{plateID}{Plate ID for the current plate.}
}
\value{
The updated master list object with mzML file paths.
}
\description{
This function restructures the directory by moving raw_data and mzML files to correct locations.
}
\examples{
\dontrun{
master_list <- msConvertR_restructure_directory(output_directory, "plateID")
}
}
