% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_set_working_directory}
\alias{msConvertR_set_working_directory}
\title{msConvertR_set_working_directory}
\usage{
msConvertR_set_working_directory(output_directory)
}
\arguments{
\item{output_directory}{Directory path for the project folder.}
}
\value{
None. The function sets the working directory.
}
\description{
This function sets the working directory to the project directory.
}
\examples{
\dontrun{
msConvertR_set_working_directory("path/to/output_directory")
}
}
