% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_setup_project_directories}
\alias{msConvertR_setup_project_directories}
\title{msConvertR_setup_project_directories}
\usage{
msConvertR_setup_project_directories(output_directory, plateID)
}
\arguments{
\item{output_directory}{Output directory for mzml files.}

\item{plateID}{Plate ID for the current plate.}
}
\value{
None. The function sets up directories.
}
\description{
This function sets up project directories for each plate ID.
}
\examples{
\dontrun{
msConvertR_setup_project_directories(output_directory, plateID)
}
}
