% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{mzml_conversion}
\alias{mzml_conversion}
\title{mzml_conversion}
\usage{
mzml_conversion(plateID, master_list)
}
\arguments{
\item{plateID}{Plate ID for the current plate.}

\item{master_list}{Master list generated internally.}
}
\value{
The updated \code{master_list} object with the mzML conversion details.
}
\description{
This function converts raw files to mzML format using ProteoWizard's msconvert tool, restructures directories, and updates the script log.
}
\examples{
\dontrun{
mzml_conversion("plateID", master_list)
}
}
