% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{optimise_retention_times}
\alias{optimise_retention_times}
\title{optimise_retention_times}
\usage{
optimise_retention_times(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{A vector of plate indices to optimise retention times for.}
}
\value{
A list containing the optimised retention times and updated MRM guide for each plate.
}
\description{
This function optimises retention times for each plate in the master list using the mzR_mrm_findR function and updates the MRM guide.
}
\examples{
\dontrun{
optimise_retention_times(master_list, plate_idx)
}
}
\keyword{internal}
