% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{PeakForgeR_import_external_functions}
\alias{PeakForgeR_import_external_functions}
\title{Utils_Global.R
Import specific functions from packages}
\description{
Utils_Global.R
Import specific functions from packages
}
\keyword{internal}
