% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{plot_control_chart}
\alias{plot_control_chart}
\title{Plot Control Chart}
\usage{
plot_control_chart(master_list, metabolite, plate_boundaries, annotate_label)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{metabolite}{The metabolite to plot in the control chart.}

\item{plate_boundaries}{A vector of plate boundaries for vertical lines in the plot.}

\item{annotate_label}{A tibble containing annotations for the plates.}
}
\value{
A \code{ggplot} object representing the control chart for the specified metabolite.
}
\description{
This function generates a control chart for a specific metabolite in the \code{master_list}.
It combines data from peak area, SIL peak area, concentration, and stat target concentration, and plots the values against the sample run index.
}
