% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{prepare_statTarget_files}
\alias{prepare_statTarget_files}
\title{Prepare statTarget Files}
\usage{
prepare_statTarget_files(FUNC_list)
}
\arguments{
\item{FUNC_list}{A list containing the project directory, master data, and metabolite list.}
}
\value{
The updated \code{FUNC_list} with created phenotype and profile files.
}
\description{
This function prepares the phenotype and profile files required for \code{statTarget} batch correction.
It creates a phenotype file with sample metadata and a profile file with metabolite data.
}
\keyword{internal}
