% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{process_plates}
\alias{process_plates}
\title{process_plates}
\usage{
process_plates(master_list, mzml_filelist)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{mzml_filelist}{A list of mzML files for each plate.}
}
\value{
The updated \code{master_list} object with processed mzML data.
}
\description{
This function processes mzML files for each plate in the master list, extracting relevant information and updating the master list.
}
\examples{
\dontrun{
process_plates(master_list, mzml_filelist)
}
}
\keyword{internal}
