% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{process_sil_target}
\alias{process_sil_target}
\title{Process SIL Target}
\usage{
process_sil_target(master_list, plate_id, data_type, template_version, sil)
}
\arguments{
\item{master_list}{A list containing project details, peak area data, and SIL templates.}

\item{plate_id}{The ID of the plate to process.}

\item{data_type}{The type of data to process (either "sorted" or "imputed").}

\item{template_version}{The version of the template to use for processing.}

\item{sil}{The name of the SIL target to process.}
}
\value{
The updated master list with calculated response and concentration data for the specified SIL target.
}
\description{
This function processes a specific SIL target by calculating the response ratio and concentration for each sample.
It retrieves the precursor names from the SIL guide, calculates the response ratio by dividing the peak area by the SIL value, and computes the concentration using the concentration factor from the template.
}
\keyword{internal}
