% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR.R
\name{qcCheckR}
\alias{qcCheckR}
\title{Quality Control Check for R}
\usage{
qcCheckR(project_directory, mrm_template_list, sample_file)
}
\arguments{
\item{project_directory}{A character string specifying the path to the project directory.}

\item{mrm_template_list}{A list of MRM templates. Default is an empty list.}
}
\value{
A list containing the processed data and generated reports.
}
\description{
This function performs a series of quality control checks on the data within a specified project directory.
}
\examples{
mrm_template_list <- list(
    v1 = list(
       SIL_guide = "templates/LGW_lipid_mrm_template_v1.csv",
       conc_guide = "templates/LGW_SIL_batch_103.csv"
      ),
    v2 = list(
       SIL_guide = "templates/LGW_lipid_mrm_template_v2.csv",
       conc_guide = "templates/LGW_SIL_batch_Ultimate_2023-03-06.csv"
      )
  )
sample_file <-  data.frame(sampleID,sample_type)
qcCheckR(project_directory = "path/to/project_directory", mrm_template_list = mrm_template_list)
}
