% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_export_all}
\alias{qcCheckR_export_all}
\title{Export All Project Outputs}
\usage{
qcCheckR_export_all(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with exported files.
}
\description{
Exports the \code{master_list} to XLSX, HTML, and RDA formats, including summary tables, QC metrics, and processed data.
}
\examples{
\dontrun{
master_list <- qcCheckR_export_all(master_list)
}
}
