% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_export_html_report}
\alias{qcCheckR_export_html_report}
\title{Export HTML Report}
\usage{
qcCheckR_export_html_report(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the exported HTML report.
}
\description{
This function exports the \code{master_list} data to an HTML report using a specified RMarkdown template.
}
\examples{
\dontrun{
master_list <- qcCheckR_export_html_report(master_list)
}
}
