% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_export_master_list_rda}
\alias{qcCheckR_export_master_list_rda}
\title{Export Master List RDA}
\usage{
qcCheckR_export_master_list_rda(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the exported RDA file.
}
\description{
This function exports the \code{master_list} data to an RDA file.
}
\examples{
\dontrun{
master_list <- qcCheckR_export_master_list_rda(master_list)
}
}
