% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_export_xlsx_file}
\alias{qcCheckR_export_xlsx_file}
\title{Export XLSX File}
\usage{
qcCheckR_export_xlsx_file(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the exported XLSX file.
}
\description{
This function exports the \code{master_list} data to an XLSX file with multiple tabs, including user guides, QC performance, sample missing values, lipid missing values, lipid QC RSD, and various data sets.
}
\examples{
\dontrun{
master_list <- qcCheckR_export_xlsx_file(master_list)
}
}
