% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_import_skyline_reports}
\alias{qcCheckR_import_skyline_reports}
\title{qcCheckR_import_skyline_reports}
\usage{
qcCheckR_import_skyline_reports(master_list)
}
\arguments{
\item{master_list}{The master list object.}
}
\value{
The updated master list object with Skyline reports imported.
}
\description{
This function imports Skyline reports from the project directory and stores them in the master list.
}
\examples{
\dontrun{
master_list <- qcCheckR_import_skyline_reports(master_list)
}
}
\keyword{internal}
