% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_PCA}
\alias{qcCheckR_PCA}
\title{PCA Analysis and Plotting}
\usage{
qcCheckR_PCA(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with PCA models, scores, and ggplot objects.
}
\description{
Performs PCA analysis on the \code{master_list} data and generates PCA plots using \code{ggplot2}.
}
\examples{
\dontrun{
master_list <- qcCheckR_PCA(master_list)
}
}
\keyword{internal}
