% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_plot_options}
\alias{qcCheckR_plot_options}
\title{Set Plot Options}
\usage{
qcCheckR_plot_options(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with plot options set.
}
\description{
This function sets the plot color, fill, shape, and size options for the \code{master_list} data.
}
\examples{
\dontrun{
master_list <- qcCheckR_plot_options(master_list)
}
}
