% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_RSD_filter}
\alias{qcCheckR_RSD_filter}
\title{RSD Filter}
\usage{
qcCheckR_RSD_filter(master_list)
}
\arguments{
\item{master_list}{Master list from previous functions.}
}
\value{
The updated \code{master_list} with RSD filter flags.
}
\description{
Filters features per plate with a \%RSD > 30\% based on the precision of measurement in the \code{master_list} data.
Applies filtering to peakArea, concentration, and statTarget concentration data sources.
}
\examples{
\dontrun{
master_list <- qcCheckR_RSD_filter(master_list)
}
}
\keyword{internal}
