% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_run_order_plots}
\alias{qcCheckR_run_order_plots}
\title{Run Order Plots}
\usage{
qcCheckR_run_order_plots(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with run order plots.
}
\description{
This function generates run order plots for the \code{master_list} data, showing PCA scores versus run order.
}
\examples{
\dontrun{
master_list <- qcCheckR_run_order_plots(master_list)
}
}
