% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_set_qc}
\alias{qcCheckR_set_qc}
\title{Set QC for Filtering}
\usage{
qcCheckR_set_qc(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with the QC type set for filtering.
}
\description{
This function sets the QC type for filtering based on the global QC pass status in the \code{master_list} data.
}
\examples{
\dontrun{
master_list <- qcCheckR_set_qc(master_list)
}
}
