% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_setup_project}
\alias{qcCheckR_setup_project}
\title{qcCheckR Setup Project}
\usage{
qcCheckR_setup_project(project_directory)
}
\arguments{
\item{project_directory}{project directory inherited from qcCheckR.R}
}
\value{
The updated \code{master_list} object with the project setup details.
}
\description{
This function sets up the project for QC by loading necessary packages, setting up the project directory, and loading and combining plate RDA files from SkylineR.
}
\examples{
\dontrun{
qcCheckR_setup_project()
}
}
