% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_setup_project_directories}
\alias{qcCheckR_setup_project_directories}
\title{qcCheckR_setup_project_directories}
\usage{
qcCheckR_setup_project_directories(master_list)
}
\arguments{
\item{master_list}{The master list object.}
}
\value{
The updated master list object with project directories set up.
}
\description{
This function sets up the project directories for the master list.
}
\examples{
\dontrun{
master_list <- qcCheckR_setup_project_directories(master_list)
}
}
\keyword{internal}
