% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_sil_IntStd_filter}
\alias{qcCheckR_sil_IntStd_filter}
\title{SIL Internal Standard Filter}
\usage{
qcCheckR_sil_IntStd_filter(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with SIL internal standard filter flags.
}
\description{
This function filters SIL internal standards based on missing values in the \code{master_list} data.
}
\examples{
\dontrun{
master_list <- qcCheckR_sil_IntStd_filter(master_list)
}
}
