% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_statTarget_batch_correction}
\alias{qcCheckR_statTarget_batch_correction}
\title{Batch Correction and Signal Drift Adjustment}
\usage{
qcCheckR_statTarget_batch_correction(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with corrected data.
}
\description{
This function performs batch correction and signal drift adjustment on the data in \code{master_list} using the \code{statTarget} package.
}
\examples{
\dontrun{
master_list <- qcCheckR_statTarget_batch_correction(master_list)
}
}
