% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_summary_report}
\alias{qcCheckR_summary_report}
\title{Summary Report}
\usage{
qcCheckR_summary_report(master_list)
}
\arguments{
\item{master_list}{master_list generated by previous functions}
}
\value{
The updated \code{master_list} with the summary report.
}
\description{
This function generates a summary report for the \code{master_list} data, including metrics for cohorts, matrix types, sample counts, lipid targets, SIL versions, missing value filter flags, and RSD percentages.
}
\examples{
\dontrun{
master_list <- qcCheckR_summary_report()
}
}
