% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_target_control_charts}
\alias{qcCheckR_target_control_charts}
\title{Target Control Charts}
\usage{
qcCheckR_target_control_charts(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with control charts.
}
\description{
This function generates control charts for the \code{master_list} data, showing the values of target metabolites and SIL internal standards across different sample types and data sources.
}
\examples{
\dontrun{
master_list <- qcCheckR_target_control_charts(master_list)
}
}
