% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_transpose_data}
\alias{qcCheckR_transpose_data}
\title{Transpose PeakForgeR Report Data}
\usage{
qcCheckR_transpose_data(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} object with transposed peak area data.
}
\description{
This function transposes PeakForgeR report data for each plate in the master list.
It reshapes the data, cleans sample names, converts values to numeric, and stores the result.
}
\examples{
\dontrun{
qcCheckR_transpose_data(master_list)
}
}
\keyword{internal}
