% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{read_mrm_guides}
\alias{read_mrm_guides}
\title{read_mrm_guides}
\usage{
read_mrm_guides(master_list, mrm_template_list)
}
\arguments{
\item{master_list}{The master list object.}

\item{mrm_template_list}{List of MRM guide file paths.}
}
\value{
The updated master list object with MRM guides.
}
\description{
This function reads MRM guides from user-supplied paths in the mrm_template_list.
Capable of reading .tsv or .csv files
}
\examples{
\dontrun{
master_list <- read_mrm_guides(master_list, mrm_template_list)
}
}
\keyword{internal}
