% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{reimport_PeakForgeR_file}
\alias{reimport_PeakForgeR_file}
\title{reimport_PeakForgeR_file}
\usage{
reimport_PeakForgeR_file(master_list, plate_idx)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{plate_idx}{The index of the plate to reimport the Skyline file for.}
}
\value{
A data frame containing the reimported PeakForgeR data with cleaned column names.
}
\description{
This function reimports a PeakForgeR file for a given plate, converts specific columns to numeric, and cleans the column names.
}
\examples{
\dontrun{
reimport_PeakForgeR_file(master_list, plate_idx)
}
}
\keyword{internal}
