% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{replace_precursor_symbols}
\alias{replace_precursor_symbols}
\title{replace_precursor_symbols}
\usage{
replace_precursor_symbols(mrm_template, columns = c("Precursor Name", "Note"))
}
\arguments{
\item{mrm_template}{dataframe of transitions (mrm_template) for SkylineR or qcCheckR}
}
\value{
Updated mrm_template with special characters replaced in 'Precursor Name' and 'Note', while the original names are preserved for the columns in original_col
}
\description{
This function replaces forward or backwards slashes in 'Precursor Name' while preserving the original naming convention
This is due to skyline cmd being unable to handle the special character
}
\examples{
\dontrun{
replace_precursor_symbols(mrm_template, columns = c("Precursor Name", "Note"))
}
}
\keyword{internal}
