% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{replace_problematic_values}
\alias{replace_problematic_values}
\title{Replace Problematic Values}
\usage{
replace_problematic_values(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A matrix with zeros replaced by NA, and infinite and NaN values also replaced by NA.
}
\description{
This function replaces zeros, infinite, and NaN values in a matrix with NA.
}
\keyword{internal}
