% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{restructure_directory}
\alias{restructure_directory}
\title{restructure_directory}
\usage{
restructure_directory(master_list, plateID)
}
\arguments{
\item{master_list}{The master list object.}

\item{plateID}{Plate ID for the current plate.}
}
\value{
The updated master list object with mzML file paths.
}
\description{
This function restructures the directory by moving wiff and mzML files to correct locations.
}
\examples{
\dontrun{
master_list <- restructure_directory(master_list, "plateID")
}
}
