% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_lipid_explorer}
\alias{run_lipid_explorer}
\title{Run Lipid Explorer
This is the main execution flow.}
\usage{
run_lipid_explorer(project_directory)
}
\arguments{
\item{project_directory}{This is the working project folder that...}
}
\value{
Report document
}
\description{
Run Lipid Explorer
This is the main execution flow.
}
