% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{run_pca_model}
\alias{run_pca_model}
\title{Run PCA Model}
\usage{
run_pca_model(master_list, source, preprocessed = FALSE)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{source}{The data source to run PCA on (e.g., "peakArea", "concentration", "concentration.statTarget").}

\item{preprocessed}{Logical indicating whether to preprocess the data (default is FALSE).}
}
\value{
The updated \code{master_list} with PCA models and scores.
}
\description{
This function runs a PCA model on the specified data source from the \code{master_list}.
It preprocesses the data by filtering out failed samples and high RSD lipids if \code{preprocessed} is set to TRUE.
It then performs PCA using the \code{ropls} package and stores the model and scores in the \code{master_list}.
}
\keyword{internal}
