% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{run_statTarget_shiftCor}
\alias{run_statTarget_shiftCor}
\title{Run statTarget shiftCor}
\usage{
run_statTarget_shiftCor(FUNC_list, master_list)
}
\arguments{
\item{FUNC_list}{A list containing the project directory, phenotype file, and profile file.}
}
\value{
The updated \code{FUNC_list} with corrected data and adjusted QC means.
}
\description{
This function runs the \code{statTarget::shiftCor} function to perform signal drift correction on the prepared phenotype and profile files.
It specifies the correction parameters and reads the corrected data from the output file.
}
