% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{run_system_command}
\alias{run_system_command}
\title{run_system_command}
\usage{
run_system_command(skyline_command, output_file)
}
\arguments{
\item{skyline_command}{developed skyline docker command}

\item{output_file}{optional path to save the command output}
}
\description{
This function wraps the system command to run docker skyline allowing for
unit testing. It captures all output (stdout and stderr) and writes it to a .txt file.
}
\examples{
\dontrun{
run_system_command("docker run skyline", "output.txt")
}
}
\keyword{internal}
