% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{set_project_details}
\alias{set_project_details}
\title{set_project_details}
\usage{
set_project_details(master_list, project_directory, plateID, QC_sample_label)
}
\arguments{
\item{master_list}{The master list object.}

\item{project_directory}{Directory path for the project folder.}

\item{plateID}{Plate ID for the current plate.}

\item{QC_sample_label}{Key for filtering QC samples from sample list.}
}
\value{
The updated master list object with project details.
}
\description{
This function sets project details in the master list.
}
\examples{
\dontrun{
master_list <- set_project_details(master_list, "path/to/project_directory", "plateID", "QC_sample_label")
}
}
\keyword{internal}
