% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{set_project_qc_type}
\alias{set_project_qc_type}
\title{Set Project QC Type}
\usage{
set_project_qc_type(master_list)
}
\arguments{
\item{master_list}{A list containing project details and QC pass status.}
}
\value{
The updated master list with the project QC type set.
}
\description{
This function sets the QC type for the project based on user-specified QC types and global QC pass status.
}
\keyword{internal}
