% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{set_working_directory}
\alias{set_working_directory}
\title{set_working_directory}
\usage{
set_working_directory(project_directory)
}
\arguments{
\item{project_directory}{Directory path for the project folder.}
}
\value{
None. The function sets the working directory.
}
\description{
This function sets the working directory to the project directory.
}
\examples{
\dontrun{
set_working_directory("path/to/project_directory")
}
}
