% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{setup_project_directories}
\alias{setup_project_directories}
\title{setup_project_directories}
\usage{
setup_project_directories(master_list)
}
\arguments{
\item{master_list}{The master list object.}
}
\value{
None. The function sets up directories.
}
\description{
This function sets up project directories for each plate ID.
}
\examples{
\dontrun{
setup_project_directories(master_list)
}
}
