% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{skyline_setup_project}
\alias{skyline_setup_project}
\title{skyline_setup_project}
\usage{
skyline_setup_project(
  project_directory,
  plateID,
  mrm_template_list,
  QC_sample_label
)
}
\arguments{
\item{project_directory}{Directory path for the project folder containing the wiff folder and wiff files.}

\item{plateID}{Plate ID for the current plate.}

\item{mrm_template_list}{List of MRM guides.}

\item{QC_sample_label}{Key for filtering QC samples from sample list.}
}
\value{
The updated \code{master_list} object with the project setup details.
}
\description{
This function sets up the project by initialising the master list, setting up project directories, and updating the script log.
}
\examples{
\dontrun{
skyline_setup_project("path/to/project_directory", "plateID", mrm_template_list, "QC_sample_label")
}
}
