% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR.R
\name{SkylineR}
\alias{SkylineR}
\title{SkylineR}
\usage{
SkylineR(project_directory, mrm_template_list = list())
}
\arguments{
\item{project_directory}{User's path to project directory containing a folder named 'wiff' with selected .wiff and .wiff.scan files.}

\item{mrm_template_list}{User's path to desired MRM templates, must be in specified format. Can contain more than one template for multi-method projects.}
}
\value{
Curated project directory containing .wiff and .mzml files, and Skyline exports.
}
\description{
This function processes \code{.wiff} files in a specified project directory and converts them to \code{.mzml} files, followed by various processing steps.
}
\examples{
SkylineR(project_directory = "USER/PATH/TO/PROJECT/DIRECTORY", mrm_template_list = list("USER/PATH/TO/user_mrm_template1.csv", "USER/PATH/TO/user_mrm_template2.csv"))
}
