% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{sort_and_filter_data}
\alias{sort_and_filter_data}
\title{Sort and Filter Data}
\usage{
sort_and_filter_data(run_order, transposed_data, sample_tags)
}
\arguments{
\item{run_order}{A data frame containing the run order information.}

\item{transposed_data}{A data frame containing the transposed peak area data.}

\item{sample_tags}{A character vector of sample tags to filter by.}
}
\value{
A data frame containing the sorted and filtered data.
}
\description{
This function sorts the run order data and filters it based on sample types.
}
