% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{stop_with_qc_error}
\alias{stop_with_qc_error}
\title{Stop with QC Error}
\usage{
stop_with_qc_error(project_name, global_qc_pass, plate_qc_passed)
}
\arguments{
\item{project_name}{A string containing the project name.}

\item{global_qc_pass}{A list containing the global QC pass status.}

\item{plate_qc_passed}{A list containing the plate QC assessment.}
}
\value{
Stops the script execution with an error message.
}
\description{
This function stops the script execution and prints a detailed error message if no viable QC type is found.
It includes information about the global QC pass status and plate QC assessment.
}
\keyword{internal}
