% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{store_environment_details}
\alias{store_environment_details}
\title{store_environment_details}
\usage{
store_environment_details(master_list)
}
\arguments{
\item{master_list}{The master list object.}
}
\value{
The updated master list object with environment details.
}
\description{
This function stores environment details in the master list.
}
\examples{
\dontrun{
master_list <- store_environment_details(master_list)
}
}
