% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionCheckR.R
\name{transition_checkR}
\alias{transition_checkR}
\title{transition_checkR}
\usage{
transition_checkR(transition_df)
}
\arguments{
\item{transition_df}{df of transitions (mrm_template) for SkylineR or qcCheckR}
}
\value{
A message if all transitions are unique or a list of non unique transitions to be fixed
}
\description{
This function checks Q1 and Q3 transitions to ensure all transitions are unique.
}
\examples{
\dontrun{
transition_checkR(transition_df)
}
}
