% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{transpose_and_merge_corrected}
\alias{transpose_and_merge_corrected}
\title{Transpose and Merge Corrected Data
This function transposes the corrected data and merges it with the metabolite list from the profile file.
It renames columns, filters out unwanted rows, and adds sample metadata.}
\usage{
transpose_and_merge_corrected(FUNC_list)
}
\arguments{
\item{FUNC_list}{A list containing the corrected data and profile file.}
}
\value{
The updated \code{FUNC_list} with transposed and merged corrected data.
}
\description{
Transpose and Merge Corrected Data
This function transposes the corrected data and merges it with the metabolite list from the profile file.
It renames columns, filters out unwanted rows, and adds sample metadata.
}
\keyword{internal}
