% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{transpose_plate_data}
\alias{transpose_plate_data}
\title{Transpose and Clean Plate Data}
\usage{
transpose_plate_data(data)
}
\arguments{
\item{data}{The Skyline report data frame for a specific plate.}
}
\value{
A transposed tibble with sample names as rows and molecule names as columns.
}
\description{
This function transposes the plate data from the Skyline report, reshaping it into a wide format.
It cleans the sample names by removing the file extension and converts area values to numeric.
}
