% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{update_sample_list}
\alias{update_sample_list}
\title{update_sample_list}
\usage{
update_sample_list(master_list, idx_plate)
}
\arguments{
\item{master_list}{A list containing project details and data.}

\item{idx_plate}{The index of the plate to update the sample list for.}
}
\value{
The updated sample list for the specified plate.
}
\description{
This function updates the sample list for a given plate in the master list.
}
\examples{
\dontrun{
update_sample_list(master_list, idx_plate)
}
}
