% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{update_script_log}
\alias{update_script_log}
\title{Update Script Log}
\usage{
update_script_log(
  master_list,
  section_name,
  previous_section_name,
  next_section_name
)
}
\arguments{
\item{master_list}{A list containing project details and script log information.}

\item{section_name}{A string representing the name of the current section.}

\item{previous_section_name}{A string representing the name of the previous section.}

\item{next_section_name}{A string representing the name of the next section.}
}
\value{
The updated \code{master_list} object with the new log information.
}
\description{
This function updates the script log in the \code{master_list} object by capturing the current time, calculating the runtime for the current section, and creating a message for the log.
}
\examples{
\dontrun{
master_list <- list(project_details = list(script_log = list(timestamps = list(start_time = Sys.time()),
                    runtimes = list(), messages = list())))
update_script_log(master_list, "section_1", "start_time", "section_2")
}
}
