% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_file_types}
\alias{validate_file_types}
\title{Validate Raw files}
\usage{
validate_file_types(input_directory)
}
\arguments{
\item{input_directory}{directory path for vendor file locations}
}
\value{
validated paths and returns message on outcome of check.
}
\description{
This function checks project directories contains wiff files and associated wiff.scan files.
}
\examples{
\dontrun{
all_file_paths <- validate_file_types(input_directory)
}
}
