% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{validate_input_directory}
\alias{validate_input_directory}
\title{Validate input Directory}
\usage{
validate_input_directory(input_directory)
}
\arguments{
\item{input_directory}{A character string representing the path to the project directory.}
}
\value{
TRUE if the validation is successful, otherwise an error is thrown.
}
\description{
This function checks if the \code{input_directory} parameter is a single string and if the specified directory exists.
}
\examples{
\dontrun{
validate_project_directory("input_directory")
}
}
