% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{validate_master_list}
\alias{validate_master_list}
\title{Validate Master List Structure}
\usage{
validate_master_list(master_list)
}
\arguments{
\item{master_list}{The master list object to validate.}
}
\value{
NULL if the structure is valid, otherwise throws an error.
}
\description{
This function checks if the master list has the required structure and data.
}
\keyword{internal}
