% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_master_list_project_directory}
\alias{validate_master_list_project_directory}
\title{validate_master_list_project_directory
This function validates the existence of the project directory specified in the master list.}
\usage{
validate_master_list_project_directory(master_list)
}
\arguments{
\item{master_list}{A list containing project details.}
}
\value{
Stops execution if the project directory does not exist.
}
\description{
validate_master_list_project_directory
This function validates the existence of the project directory specified in the master list.
}
\examples{
\dontrun{
validate_master_list_project_directory(master_list)
}
}
\keyword{internal}
