% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_mrm_template_list}
\alias{validate_mrm_template_list}
\title{Validate MRM Template List}
\usage{
validate_mrm_template_list(mrm_template_list)
}
\arguments{
\item{mrm_template_list}{A list of character strings representing the paths to MRM templates.}
}
\value{
TRUE if the validation is successful, otherwise an error is thrown.
}
\description{
This function checks if the \code{mrm_template_list} parameter is a list of strings.
}
\examples{
\dontrun{
validate_mrm_template_list(list("path/to/template1.csv", "path/to/template2.csv"))
}
}
