% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_project_directory}
\alias{validate_project_directory}
\title{Validate Project Directory}
\usage{
validate_project_directory(project_directory)
}
\arguments{
\item{project_directory}{A character string representing the path to the project directory.}
}
\value{
TRUE if the validation is successful, otherwise an error is thrown.
}
\description{
This function checks if the \code{project_directory} parameter is a single string and if the specified directory exists.
}
\examples{
\dontrun{
validate_project_directory("path/to/project_directory")
}
}
