% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_proteowizard_skyline}
\alias{validate_proteowizard_skyline}
\title{Validate Proteowizard and Skyline install}
\usage{
validate_proteowizard_skyline()
}
\value{
Returns a program found if installed in correct location or stops script and asks user to install if not.
}
\description{
This function check proteowizard and skyline are correctly installed in users C dive program files.
}
\examples{
\dontrun{
validate_proteowizard_skyline()
}
}
