% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{validate_qc_types}
\alias{validate_qc_types}
\title{Validate QC Types}
\usage{
validate_qc_types(run_order, sample_tags)
}
\arguments{
\item{run_order}{A data frame containing the run order information.}

\item{sample_tags}{A character vector of QC types to validate against.}
}
\value{
NULL if validation passes, otherwise throws an error.
}
\description{
This function validates the sample types in the run order against the provided sample tags.
}
\keyword{internal}
