% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_qcCheckR_mrm_template_list}
\alias{validate_qcCheckR_mrm_template_list}
\title{Validate qcCheckR mrm template list}
\usage{
validate_qcCheckR_mrm_template_list(master_list)
}
\arguments{
\item{mrm_template_list}{A list specifying the file paths for the templates.}
}
\value{
TRUE if validation passes. Stops execution if validation fails.
}
\description{
This function validates the mrm_template_list list by checking the column headers and ensuring there are no NA or NULL values in the SIL_guide and conc_guide files.
}
\examples{
\dontrun{
validate_qcCheckR_mrm_template_list(master_list)
}
}
