% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_template_info}
\alias{validate_template_info}
\title{Validate mrm template list}
\usage{
validate_template_info(mrm_template_list)
}
\arguments{
\item{mrm_template_list}{A list specifying the file paths for the templates.}
}
\value{
NULL. Stops execution if validation fails.
}
\description{
This function validates the mrm_template_list list by checking the column headers and ensuring there are no NA or NULL values in the SIL_guide and conc_guide files.
}
\examples{
validate_template_info(mrm_template_list)
}
