% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_wiff_file}
\alias{validate_wiff_file}
\title{Validate Wiff files}
\usage{
validate_wiff_file(project_directory)
}
\value{
validated paths and returns message on outcome of check.
}
\description{
This function checks project directorie contains wiff files and associated wiff.scan files.
}
\examples{
\dontrun{
wiff_file_paths <- validate_wiff_file()
}
}
