% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakForgeR_Utils.R
\name{version_selector}
\alias{version_selector}
\title{version_selector}
\usage{
version_selector(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
Updated version list with the most probabilistic version first.
}
\description{
ANPC specific helper function to reorder first version tried based on jwist naming conventions.
}
\examples{
\dontrun{
version_selector(master_list)
}
}
\keyword{internal}
