% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineR_Utils.R
\name{wait_until_files_free}
\alias{wait_until_files_free}
\title{wait_until_files_free
This function waits until files are not in use by attempting to rename them.}
\usage{
wait_until_files_free(files_to_copy, max_wait = 60, max_retries = 30)
}
\arguments{
\item{files_to_copy}{A vector of file paths to check.}

\item{max_wait}{Maximum wait time (in seconds) for the system prior to deleting the moved folder. Default is 60}

\item{max_retries}{Maximum number of attempts to try move/delete prior to error. Default is 30.}
}
\value{
None. The function waits until files are free.
}
\description{
wait_until_files_free
This function waits until files are not in use by attempting to rename them.
}
\examples{
\dontrun{
wait_until_files_free(files_to_copy, max_wait = 60, max_retries = 30)
}
}
