% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{determine_qc_type}
\alias{determine_qc_type}
\title{Determine QC Type}
\usage{
determine_qc_type(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
A string indicating the QC type ("pqc", "ltr", or "unknown").
}
\description{
This function determines the QC type based on the global QC pass status.
It checks if there are multiple QC types that have passed.
Assesses if a secondary QC sample is available such as a pooled Quality controls.
}
\keyword{internal}
