% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{move_folder}
\alias{move_folder}
\title{Move Folder}
\usage{
move_folder(source_dir, dest_dir, max_wait = 60, max_retries = 30)
}
\arguments{
\item{source_dir}{Directory path for the folder to copy.}

\item{dest_dir}{Directory path for the folder to be copied to.}

\item{max_wait}{Maximum wait time (in seconds) for the system prior to deleting the moved folder.}
}
\value{
None. The function performs the move operation and prints a message upon successful completion.
}
\description{
This function moves a folder from the source directory to the destination directory, waits until files are not in use, and then deletes the source directory.
}
\examples{
\dontrun{
move_folder("path/to/source", "path/to/destination", 60)
}
}
