% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR.R
\name{msConvertR}
\alias{msConvertR}
\title{This function is used to convert vendor files to \code{.mzml} files
using the \code{msconvert} tool from ProteoWizard.}
\usage{
msConvertR(input_directory, output_directory)
}
\arguments{
\item{input_directory}{The directory where the vendor files are located.}

\item{output_directory}{The directory where the converted \code{.mzml} files will be saved.}
}
\value{
None. The function performs the conversion and restructures directories.
}
\description{
This function is used to convert vendor files to \code{.mzml} files
using the \code{msconvert} tool from ProteoWizard.
}
\examples{
\dontrun{
# example code
 msConvertR(input_directory = "path/to/input_directory", output_directory = "path/to/output_directory")
 }
}
