% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msConvertR_Utils.R
\name{msConvertR_mzml_conversion}
\alias{msConvertR_mzml_conversion}
\title{msConvertR_mzml_conversion}
\usage{
msConvertR_mzml_conversion(
  input_directory,
  output_directory,
  file_path,
  plateID
)
}
\arguments{
\item{input_directory}{Directory path for vendor files.}

\item{output_directory}{Directory path for the mzml output.}

\item{file_path}{Path to the vendor file to be converted.}

\item{plateID}{Plate ID for the current plate.}
}
\value{
Converted mzml files.
}
\description{
This function converts raw vendor files to mzML format using ProteoWizard's msconvert tool, restructures directories, and updates the script log.
}
\examples{
\dontrun{
msConvertR_mzml_conversion(input_directory, output_directory, file_path, plateID)
}
}
