% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR.R
\name{qcCheckR}
\alias{qcCheckR}
\title{Quality Control Check for R}
\usage{
qcCheckR(
  user_name,
  project_directory,
  mrm_template_list = NULL,
  QC_sample_label = NULL,
  sample_tags = NULL,
  mv_threshold = 50
)
}
\arguments{
\item{user_name}{A character string specifying the name of the user running the QC checks.}

\item{project_directory}{A character string specifying the path to the project directory.}

\item{mrm_template_list}{A list of MRM templates and associated concentration guide. Must have specific column names. See examples for structure of mrm_template_list. Must include mrm_guide labelled as "SIL_guide" and associated concentration guide labelled as "conc_guide". Can contain multiple combinations stored as separate lists, see examples.}

\item{QC_sample_label}{A character string containing the key tags to filter QC samples from file names.  E.g. "qc".}

\item{sample_tags}{A character vector specifying the tags to filter sample types from file names. E.g. c("sample","control","blank", "qc").}

\item{mv_threshold}{A numeric value  between 0 and 100 specifying the threshold for missing values in the data. Default is 0.5 (50\%).}
}
\value{
A list containing the processed data and generated reports.
}
\description{
This function performs a series of quality control checks on the data within a specified project directory.
If you have not used the SkylineR function to generate reports please ensure your report file names contain "\emph{skylineR_1}" to ensure the function can correctly identify the files in your project directory.
}
\examples{
\dontrun{

library(MetaboExploreR)

#Load example mrm_template_list
  file_path <- system.file("extdata", "LGW_lipid_mrm_template_v1.tsv", package = "MetaboExploreR")
  sample_metadata_example <- read_tsv(file_path)

#Load example conc_guide
  file_path <- system.file("extdata", "LGW_SIL_batch_Ultimate_2023_03_06.tsv", package = "MetaboExploreR")
  sample_metadata_example <- read_tsv(file_path)

#Load example report file
  file_path <- system.file("extdata", "Example_xskylineR_report.csv", package = "MetaboExploreR")
  report_file <- read.csv(file_path)

#Run qcCheckR function
qcCheckR(user_name = "user1",
         project_directory = "path/to/project_directory",
         mrm_template_list = list(v1 = list(
                                    SIL_guide = path to/mrm_guide1.tsv,
                                    conc_guide = path to/SIL_concentration_guide1.tsv),
                                  v2 = list(
                                    SIL_guide = path to/mrm_guide2.tsv,
                                    conc_guide = path to/SIL_concentration_guide2.tsv)
                                 ),
         QC_sample_label = "qc",
         sample_tags = c("sample","control","blank", "qc"),
         mv_threshold = 0.5) #default is 50\% missing values
}
}
