% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_sample_filter}
\alias{qcCheckR_sample_filter}
\title{Sample Filter}
\usage{
qcCheckR_sample_filter(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with sample filter flags.
}
\description{
This function flags samples based on missing values and summed signal intensity in the \code{master_list} data.
}
\examples{
\dontrun{
master_list <- qcCheckR_sample_filter(master_list)
}
}
