% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{qcCheckR_sort_data}
\alias{qcCheckR_sort_data}
\title{Sort and QC Check Data}
\usage{
qcCheckR_sort_data(master_list)
}
\arguments{
\item{master_list}{A list containing project details and data.}
}
\value{
The updated \code{master_list} with sorted data and QC check results.
}
\description{
This function sorts the data in \code{master_list} by run order and performs QC checks.
}
\examples{
\dontrun{
master_list <- qcCheckR_sort_data(master_list)
}
}
