% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcCheckR_Utils.R
\name{qcCheckR_transpose_data}
\alias{qcCheckR_transpose_data}
\title{qcCheckeR Transpose Data}
\usage{
qcCheckR_transpose_data(master_list)
}
\arguments{
\item{master_list}{Master list containing peak area data.}
}
\value{
The updated \code{master_list} object with transposed peak area data.
}
\description{
This function transposes peak area data for each plate in the master list, converting it to a wide format and storing the result.
}
\examples{
\dontrun{
qcCheckR_transpose_data(master_list)
}
}
