% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_mrm_template_list}
\alias{validate_mrm_template_list}
\title{Validate MRM Template List}
\usage{
validate_mrm_template_list(mrm_template_list, user_name)
}
\arguments{
\item{mrm_template_list}{A list of character strings or a named list of data frames representing MRM templates.}

\item{user_name}{A character string identifying the user.}
}
\value{
NULL or an ANPC mrm_template_list if mrm_template_list is NULL and user_name is ANPC
}
\description{
This function checks if the \code{mrm_template_list} parameter is valid and contains required columns.
}
\examples{
\dontrun{
validate_mrm_template_list(list("path/to/template1.csv", "path/to/template2.csv"), "user")
}
}
\keyword{internal}
